changes based on request.Renamed functions in the context file and deleted the userSearch file
